refactor(modals): extract common Modal wrapping to ModalWrapper
